backend/requirements.txt
Django==5.0.6
requests==2.32.3
django-cors-headers==4.3.1
python-dotenv==1.0.1
